+++ src/api/albums/entities/album.entity.ts
