+++ src/api/artists/artists.module.ts
