+++ src/api/artists/artists.service.ts
