+++ src/api/artists/dto/create-artist.dto.ts
