+++ src/api/artists/dto/update-artist.dto.ts
