+++ src/api/artists/entities/artist.entity.ts
