+++ src/api/favorites/decorators/index.ts
