+++ src/api/favorites/dto/favorites-response.dto.ts
