+++ src/api/favorites/entities/favorite-album.entity.ts
