+++ src/api/favorites/entities/favorite-artist.entity.ts
