+++ src/api/favorites/favorites.module.ts
