+++ src/api/favorites/interfaces/favorite.interface.ts
