+++ src/api/tracks/tracks.controller.ts
@@ -12,17 +12,17 @@ import {
 import { TracksService } from './tracks.service';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
-import { IdNotFoundError } from '../shared/errors/id-not-found.error';
+import { IdNotFoundError } from '../../shared/errors/id-not-found.error';
 import { ApiTags } from '@nestjs/swagger';
 import { Track } from './entities/track.entity';
-import { ParseUUIDV4Pipe } from '../shared/pipes/parse-uuid-v4.pipe';
+import { ParseUUIDV4Pipe } from '../../shared/pipes/parse-uuid-v4.pipe';
 import {
   ApiCreate,
   ApiFindAll,
   ApiFind,
   ApiUpdate,
   ApiDelete,
-} from '../shared/decorators';
+} from '../../shared/decorators';
 
 @ApiTags('Tracks')
 @Controller('track')
